% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymon-print.R
\name{format.ymon}
\alias{format.ymon}
\title{Format a ymon}
\usage{
\method{format}{ymon}(x, ...)
}
\arguments{
\item{x}{A ymon.}

\item{...}{Unused.}
}
\description{
The ymon \code{format()} method is purposefully implemented without a \verb{format =}
argument, and always prints as \code{YYYY-MM}. To use a custom format, the ymon
would have to be converted to POSIXlt, which is a very slow operation. It is
also buggy when outside the year range of 0-9999 because strptime doesn't
support it. If you want to use a custom format, convert to Date first with
\code{\link[=as.Date]{as.Date()}}, and then format that.
}
\examples{
format(ymon(2019, 01))

# NA values are printed as the actual character string
# for better behavior when combined with `print(format(x), quote = FALSE)`.
format(ymon(NA, NA))
}
