% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym.R
\name{is_ym}
\alias{is_ym}
\title{Is \code{x} a ym?}
\usage{
is_ym(x)
}
\arguments{
\item{x}{\verb{[vector]}

An object.}
}
\value{
A logical of length 1. \code{TRUE} if \code{x} is a ym, otherwise \code{FALSE}.
}
\description{
Test if \code{x} is a ym object.
}
\examples{
is_ym(new_ym(1L))
is_ym(1)
}
