% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymon.R
\name{is_ymon}
\alias{is_ymon}
\title{Is \code{x} a ymon?}
\usage{
is_ymon(x)
}
\arguments{
\item{x}{\verb{[vector]}

An object.}
}
\value{
A logical of length 1. \code{TRUE} if \code{x} is a ymon, otherwise \code{FALSE}.
}
\description{
Test if \code{x} is a ymon object.
}
\examples{
is_ymon(new_ymon(1L))
is_ymon(1)
}
