% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymon.R
\name{new_ymon}
\alias{new_ymon}
\title{Construct a new year month}
\usage{
new_ymon(x = integer())
}
\arguments{
\item{x}{\verb{[integer]}

The number of months since \code{1970-01-01}.}
}
\value{
A ymon object.
}
\description{
\code{new_ymon()} is a fast constructor of ymon objects, with minimal checking.
A ymon object is constructed from the integer number of months since the Unix
epoch of 1970-01-01. For more robust creation of ymon objects, see \code{\link[=ymon]{ymon()}} or
\code{\link[=as_ymon]{as_ymon()}}.
}
\examples{
# Internally stored as the number of months since 1970-01-01
new_ymon(0L)
new_ymon(100L)
}
