% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeclass-package.R
\docType{package}
\name{timeclass-package}
\alias{timeclass}
\alias{timeclass-package}
\title{timeclass: Extended Date Classes}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/timeclass}
  \item Report bugs at \url{https://github.com/DavisVaughan/timeclass/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
