% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym.R
\name{ym-accessors}
\alias{ym-accessors}
\alias{ym_year}
\alias{ym_month}
\title{Accessors}
\usage{
ym_year(x)

ym_month(x)
}
\arguments{
\item{x}{\verb{[ym]}

A ym object.}
}
\description{
\itemize{
\item \code{ym_year()} returns the ym year.
\item \code{ym_month()} returns the ym month.
}
}
\examples{
x <- ym(
  year = c(2000, NA, 2001),
  month = c(1, 2, 3)
)

ym_year(x)
}
