% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym.R
\name{ym}
\alias{ym}
\title{Construct a new year month}
\usage{
ym(year, month)
}
\arguments{
\item{year}{\verb{[integer]}

The year value.}

\item{month}{\verb{[integer]}

The month value, in the range of \code{1-12}.}
}
\value{
A ym object.
}
\description{
\code{ym()} is a high level constructor for a year month object. It accepts
integer year and month components to build the object from.
}
\details{
\code{year} and \code{month} are recycled to a common size.
}
\examples{
ym(2019, 2)
ym(2019, c(11, 12))
}
