% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym-parse.R
\name{ym_parse}
\alias{ym_parse}
\title{Parse a character vector as year month}
\usage{
ym_parse(x, format = "\%Y-\%m")
}
\arguments{
\item{x}{\verb{[character]}

A character vector to coerce to ym.}

\item{format}{\verb{[character(1)]}

A format to parse character input with. Should generally only consist
of format tokens related to year or month. Common formats are \code{"\%Y-\%m"},
\code{"\%b \%Y"}, and \code{"\%Y \%b"}.}
}
\description{
\code{ym_parse()} is a parser for turning character input into ym objects. The
character method for \code{as_ym()} is very strict, and only parses input of the
form \code{"YYYY-MM"}, erroring on any failures. \code{ym_parse()} is more flexible
and will only issue a warning if parsing fails.

Internally:
\itemize{
\item \code{"-\%d"} is appended to the \code{format}, and \code{"-01"} is appended to \code{x}.
\item An attempt to parse as a Date is then made.
\item The resulting Date is converted to ym, with a warning if any input failed
to parse. Failures result in \code{NA}.
}
}
\examples{
ym_parse("1970-01")
ym_parse("1970 Jan", format = "\%Y \%b")

# Unparseable input results in `NA`, with a warning
try(ym_parse(c("1970-00", "1970-01")))
}
