% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymon.R
\name{ymon-accessors}
\alias{ymon-accessors}
\alias{ymon_year}
\alias{ymon_month}
\title{Accessors}
\usage{
ymon_year(x)

ymon_month(x)
}
\arguments{
\item{x}{\verb{[ymon]}

A ymon object.}
}
\description{
\itemize{
\item \code{ymon_year()} returns the ymon year.
\item \code{ymon_month()} returns the ymon month.
}
}
\examples{
x <- ymon(
  year = c(2000, NA, 2001),
  month = c(1, 2, 3)
)

ymon_year(x)
}
