% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymon.R
\name{ymon}
\alias{ymon}
\title{Construct a new ymon}
\usage{
ymon(year, month)
}
\arguments{
\item{year}{\verb{[integer]}

The year value.}

\item{month}{\verb{[integer]}

The month value, in the range of \code{1-12}.}
}
\value{
A ymon object.
}
\description{
\code{ymon()} is a high level constructor for a ymon object. It accepts
integer year and month components to build the object from.
}
\details{
\code{year} and \code{month} are recycled to a common size.
}
\examples{
ymon(2019, 2)
ymon(2019, c(11, 12))
}
