% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymon-parse.R
\name{ymon_parse}
\alias{ymon_parse}
\title{Parse a character vector as year month}
\usage{
ymon_parse(x, format = "\%Y-\%m")
}
\arguments{
\item{x}{\verb{[character]}

A character vector to coerce to ymon.}

\item{format}{\verb{[character(1)]}

A format to parse character input with. Should generally only consist
of format tokens related to year or month. Common formats are \code{"\%Y-\%m"},
\code{"\%b \%Y"}, and \code{"\%Y \%b"}.}
}
\description{
\code{ymon_parse()} is a parser for turning character input into ymon objects. The
character method for \code{as_ymon()} is very strict, and only parses input of the
form \code{"YYYY-MM"}, erroring on any failures. \code{ymon_parse()} is more flexible
and will only issue a warning if parsing fails.

Internally:
\itemize{
\item \code{"-\%d"} is appended to the \code{format}, and \code{"-01"} is appended to \code{x}.
\item An attempt to parse as a Date is then made.
\item The resulting Date is converted to ymon, with a warning if any input failed
to parse. Failures result in \code{NA}.
}
}
\examples{
ymon_parse("1970-01")
ymon_parse("1970 Jan", format = "\%Y \%b")

# Unparseable input results in `NA`, with a warning
try(ymon_parse(c("1970-00", "1970-01")))
}
