% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym-as.R
\name{as_ym}
\alias{as_ym}
\alias{as_ym.Date}
\alias{as_ym.POSIXct}
\alias{as_ym.POSIXlt}
\alias{as_ym.integer}
\alias{as_ym.double}
\alias{as_ym.character}
\title{Coerce to year month}
\usage{
as_ym(x, ...)

\method{as_ym}{Date}(x, ...)

\method{as_ym}{POSIXct}(x, ...)

\method{as_ym}{POSIXlt}(x, ...)

\method{as_ym}{integer}(x, ...)

\method{as_ym}{double}(x, ...)

\method{as_ym}{character}(x, ...)
}
\arguments{
\item{x}{\verb{[vector]}

An object to coerce to ym.}

\item{...}{Not used.}
}
\description{
\itemize{
\item Date, POSIXct, and POSIXlt are converted directly by extracting the year
and month from \code{x}. Any day, hour, minute, or second components are
dropped. Time zone information is not retained.
\item Integer and double input are assumed to be the number of months since the
Unix origin of 1970-01-01.
\item Character input is assumed to be provided in a format containing only
information about the year and month, such as \code{"1970-01"} or \code{"Jan 1970"}.
It is parsed using the defaults of \code{\link[=ym_parse]{ym_parse()}}.
}
}
\examples{
# Extra information such as days, hours, or time zones are dropped
as_ym(as.Date("2019-05-03"))
as_ym(as.POSIXct("2019-03-04 01:01:01", tz = "America/New_York"))

# Integers are interpreted as the number of months since 1970-01-01
as_ym(0L)
as_ym(12L)
}
