% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym.R
\name{new_ym}
\alias{new_ym}
\title{Construct a new year month}
\usage{
new_ym(x = integer())
}
\arguments{
\item{x}{\verb{[integer]}

The number of months since \code{1970-01-01}.}
}
\value{
A ym object.
}
\description{
\code{new_ym()} is a fast constructor of ym objects, with minimal checking.
A ym object is constructed from the integer number of months since the Unix
epoch of 1970-01-01. For more robust creation of ym objects, see \code{\link[=ym]{ym()}} or
\code{\link[=as_ym]{as_ym()}}.
}
\examples{
# Internally stored as the number of months since 1970-01-01
new_ym(0L)
new_ym(100L)
}
