% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym-vctrs-cast.R, R/ym-vctrs-ptype.R
\name{vec_cast.ym}
\alias{vec_cast.ym}
\alias{vctrs-compat}
\alias{vec_ptype2.ym}
\title{vctrs compatibility functions}
\usage{
\method{vec_cast}{ym}(x, to, ...)

\method{vec_ptype2}{ym}(x, y, ...)
}
\arguments{
\item{x, y}{Objects.}

\item{to}{Type to cast to.}

\item{...}{Used to pass along error message information.}

\item{op}{An arithmetic operator as a string.}
}
\value{
See the corresponding vctrs function for the exact return value.
}
\description{
These functions are the extensions that allow timeclass objects to
work with vctrs.
}
