% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearmonth-vctrs-cast.R,
%   R/yearmonth-vctrs-ptype.R
\name{vec_cast.yearmonth}
\alias{vec_cast.yearmonth}
\alias{vctrs-compat}
\alias{vec_ptype2.yearmonth}
\title{vctrs compatibility functions}
\usage{
\method{vec_cast}{yearmonth}(x, to, ...)

\method{vec_ptype2}{yearmonth}(x, y, ...)
}
\arguments{
\item{x, y}{Objects.}

\item{to}{Type to cast to.}

\item{...}{Used to pass along error message information.}

\item{op}{An arithmetic operator as a string.}
}
\value{
See the corresponding vctrs function for the exact return value.
}
\description{
These functions are the extensions that allow timeclass objects to
work with vctrs.
}
