% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymon-vctrs-arith.R, R/ymon-vctrs-cast.R,
%   R/ymon-vctrs-ptype.R
\name{vec_arith.ymon}
\alias{vec_arith.ymon}
\alias{vec_cast.ymon}
\alias{vctrs-compat}
\alias{vec_ptype2.ymon}
\title{vctrs compatibility functions}
\usage{
\method{vec_arith}{ymon}(op, x, y, ...)

\method{vec_cast}{ymon}(x, to, ...)

\method{vec_ptype2}{ymon}(x, y, ...)
}
\arguments{
\item{op}{An arithmetic operator as a string.}

\item{x, y}{Objects.}

\item{...}{Used to pass along error message information.}

\item{to}{Type to cast to.}
}
\value{
See the corresponding vctrs function for the exact return value.
}
\description{
These functions are the extensions that allow datea objects to
work with vctrs.
}
