% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym.R
\name{ym}
\alias{ym}
\title{Construct a new year month}
\usage{
ym(year = NULL, month = NULL)
}
\arguments{
\item{year}{\verb{[integer / NULL]}

The year value. If left as \code{NULL} and any \code{month} values are present,
a default of \code{0} is used.}

\item{month}{\verb{[integer / NULL]}

The month value, in the range of \code{1-12}. If left as \code{NULL} and any \code{year}
values are present, a default of \code{1} is used.}
}
\value{
A ym object made from \code{year-month}.
}
\description{
\code{ym()} is a high level constructor for a year month object. It accepts
integer year and month components to build the object from.
}
\examples{
ym()

ym(2019)

ym(2019, 2)

ym(month = 2)
}
